+++ tests/Fixtures/Models/User.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ElPandaPe\FilamentActivitylog\Tests\Fixtures\Models;
 
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Spatie\Activitylog\Models\Concerns\LogsActivity;
 use Spatie\Activitylog\Support\LogOptions;
@@ -15,6 +16,12 @@ final class User extends Authenticatable
     protected $table = 'users';
 
     protected $guarded = [];
+
+    /** @return BelongsToMany<Role, $this> */
+    public function roles(): BelongsToMany
+    {
+        return $this->belongsToMany(Role::class, 'role_user');
+    }
 
     public function getActivitylogOptions(): LogOptions
     {